apps/api/requirements-embedded-report.txt
python-dotenv==1.2.1
PyPDF2==3.0.1
Pillow==12.0.0
pytesseract==0.3.10
requests==2.32.5
pdf2image==1.17.0
json-repair==0.33.0
